1.23 QueryDSL 추가(통계관련 쿼리) 한국 방문객 많은 순으로 상위 7개 나라의 정보 출력하기 (완료)

+++ src/main/java/github/tourism/data/entity/statistic/Gender_Statistic.java
@@ -26,9 +26,9 @@ public class Gender_Statistic {
     private Integer total_population;
     private Integer previous_total_population;
     private BigDecimal total_increasement;
-    private BigDecimal male_population;
+    private Integer male_population;
     private BigDecimal male_ratio;
-    private BigDecimal female_population;
+    private Integer female_population;
     private BigDecimal female_ratio;
